+++ Project2/Project2/src2/symboltable.h
@@ -2,18 +2,18 @@ class SymbolTable
 {
 public:
     SymbolTable() {}
-    void insert(string variable, double value);
-    double lookUp(string variable) const;
+    void insert(string variable, int value);
+    int lookUp(string variable) const;
 private:
     struct Symbol
     {
-        Symbol(string variable, double value)
+        Symbol(string variable, int value)
         {
             this->variable = variable;
             this->value = value;
         }
         string variable;
-        double value;
+        int value;
     };
     vector <Symbol> elements;
 };
